currency-exchange

+++ src/main/java/com/shop/PetProject/utils/converters/ProductEntityToRecordConverter.java
@@ -5,13 +5,15 @@ import com.shop.PetProject.models.ProductEntity;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class ProductEntityToRecordConverter implements Converter<ProductEntity, ProductDTO> {
     @Override
     public ProductDTO convert(ProductEntity source) {
         return ProductDTO.builder()
                 .name(source.getName())
-                .price(source.getPrice())
+                .price(BigDecimal.valueOf(source.getPrice()))
                 .description(source.getDescription())
                 .category(source.getCategory())
                 .quantity(source.getQuantity())
